backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.8
python-decouple==3.8
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-filter==23.4
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
